Move unit tests to a test/unit subfolder
